==> 02_data_statistics.Rmd <==
# Data statistics

```{r load_data_stats}
load("data/08032025data.Rdata")
sample_metadata <- sample_metadata %>%
    mutate(time_point=case_when(
    time_point %in% c("Transplant1") ~ "Donor1",
    TRUE ~ time_point
  ))%>%
    mutate(time_point=case_when(
    time_point %in% c("Transplant2") ~ "Donor2",
    TRUE ~ time_point
  )) %>% 
    mutate(Population=case_when(
    Population %in% c("Cold_wet") ~ "Cold",
    TRUE ~ Population
  ))%>% 
    mutate(Population=case_when(
    Population %in% c("Hot_dry") ~ "Warm",
    TRUE ~ Population
  ))
```

## Sequencing reads statistics

```{r reads_stats_filtering}
data_stats_filter <- data_stats %>%
  left_join(., sample_metadata[c(1,4,7,10)], by = join_by(sample == Tube_code))  %>%
  filter(Population!="NA")
```

```{r reads_stats}
data_stats_filter$raw_reads %>% sum()
data_stats_filter$raw_reads %>% mean()
data_stats_filter$raw_reads %>% sd()
```

## DNA fractions
```{r data_fractions_stats}
data_stats_filter %>%
    mutate(mapped_perc=mapped_mags/trimmed_reads) %>%
    summarise(mean=mean(mapped_perc),sd=sd(mapped_perc)) %>% 
  mutate(mean=str_c(round(mean,3),"±",round(sd,3))) %>% 
  select(mean) %>% 
  pull() %>% 
  cat()
```

```{r data_fractions_plot, message=FALSE, warning=FALSE, fig.height=5, fig.width=10, fig.fullwidth=TRUE}
data_stats_filter %>%
  mutate(
    low_quality = raw_reads - trimmed_reads,
    unmapped_reads = trimmed_reads - mapped_lizard - mapped_mags
  ) %>%
  select(sample, low_quality, mapped_lizard, mapped_mags, unmapped_reads) %>%
  pivot_longer(-sample) %>%
  mutate(name=factor(name,levels=c("low_quality","mapped_lizard","unmapped_reads","mapped_mags"))) %>%
  left_join(., sample_metadata, by = join_by(sample == Tube_code)) %>%
  filter(!is.na(time_point)) %>%
  ggplot(aes(x = sample, y = value, fill = name)) +
      geom_bar(stat = "identity", position = "fill") +
      scale_fill_manual(name="Sequence type",
                    breaks=c("low_quality","mapped_lizard","unmapped_reads","mapped_mags"),
                    labels=c("Low quality","Mapped to host","Unmapped","Mapped to MAGs"),
                    values=c("#CCCCCC", "#bcdee1", "#d8b8a3","#93655c"))+
      facet_grid(~factor(time_point, level=c("Wild", "Acclimation", "Antibiotics", "Donor1", "Donor2", "FMT1", "FMT2")), 
             scales = "free") +
      theme_minimal() +
      theme(axis.text.x = element_text(angle = 90, hjust = 1, size=0)) +
      labs(y="DNA sequence fraction",x="Samples")
```

## Recovered microbial fraction

```{r data_estimations_plot, message=FALSE, warning=FALSE, fig.height=4, fig.width=10, fig.fullwidth=TRUE}
data_stats_filter %>%
  mutate(
    unmapped_reads = trimmed_reads - mapped_lizard - mapped_mags,
    mag_proportion = mapped_mags / (mapped_mags + unmapped_reads),
    singlem_read_fraction = singlem_read_fraction
  ) %>%
  select(sample, mag_proportion, singlem_read_fraction) %>%
  mutate(
    mag_proportion = if_else(singlem_read_fraction == 0, 0, mag_proportion),
    singlem_read_fraction = if_else(singlem_read_fraction == 0, NA, singlem_read_fraction),
    singlem_read_fraction = if_else(singlem_read_fraction < mag_proportion, NA, singlem_read_fraction),
    singlem_read_fraction = if_else(singlem_read_fraction > 1, 1, singlem_read_fraction)
  ) %>%
  pivot_longer(-sample, names_to = "proportion", values_to = "value") %>%
  mutate(
    proportion = factor(
      proportion,
      levels = c("mag_proportion", "singlem_read_fraction")
    )
  ) %>%
  left_join(., sample_metadata, by = join_by(sample == Tube_code)) %>%
  filter(!is.na(time_point)) %>%
  ggplot(aes(x = sample, y = value, color = proportion)) +
      geom_line(aes(group = sample), color = "#f8a538") +
      geom_point() +
      scale_color_manual(name="Proportion",
                    breaks=c("mag_proportion","singlem_read_fraction"),
                    labels=c("Recovered","Estimated"),
                    values=c("#52e1e8", "#876b53"))+
      theme_minimal() +
      facet_grid(~factor(time_point, level=c("Wild", "Acclimation", "Antibiotics", "Donor1", "Donor2", "FMT1", "FMT2")), 
             scales = "free") +
      labs(y = "Samples", x = "Prokaryotic fraction") +
      scale_y_continuous(limits = c(0, 1)) +
      theme(
        axis.text.y = element_text(size = 4),
        axis.text.x = element_text( angle = 90, vjust = 0.5, hjust = 1, size = 0),
        legend.position = "right"
      )
```

### Domain-adjusted mapping rate (DAMR)

```{r damr}
data_stats_filter %>%
  mutate(
    unmapped_reads = trimmed_reads - mapped_lizard - mapped_mags,
    mag_proportion = mapped_mags / (mapped_mags + unmapped_reads),
    singlem_read_fraction = singlem_read_fraction
  ) %>%
  mutate(damr=pmin(1, mag_proportion/singlem_read_fraction)) %>%
  filter(!is.na(time_point)) %>%
  select(sample,damr, time_point, Population, type) %>%
  tt()
```
